I/O서비스 리팩토링
