+++ backend/rand-chat-chatAPI/src/main/java/com/rand/controller/ChatRoomWfxAPIController.java
@@ -12,6 +12,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.handler.annotation.Header;
 import org.springframework.web.bind.annotation.*;
 
+//WebClient api
 @RestController
 @RequiredArgsConstructor
 @Slf4j
@@ -19,15 +20,17 @@ import org.springframework.web.bind.annotation.*;
 public class ChatRoomWfxAPIController {
     private final ChatWfxApiService chatApiService;
 
+    //실제 참여하고있는지 검증을 위한 api
     @GetMapping
     public Boolean isRealYourRoom(@Valid  RoomValidDTO roomValidDTO){
        return chatApiService.isRealYourRoom(roomValidDTO);
     }
-
+    //상대방 정보를 조회하기위한 api
     @GetMapping("/{chatRoomId}/opsMem")
     public Members getOpsMem(@PathVariable Integer chatRoomId){
         return chatApiService.getOpsMem(chatRoomId);
     }
+    //입장플래그를 업데이트하기 위한 API
     @PutMapping("/enter/{chatRoomId}")
     public ResponseEntity<ResponseDTO<Void>> asyncEnterRoomUpdateInfo(@PathVariable Integer chatRoomId){
         return chatApiService.asyncEnterRoomUpdateInfo(
