+++ backend/rand-chat-matching/src/test/untitled/src/main/java/com/rand/Main.java
@@ -1,0 +1,7 @@
+package com.rand;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
