+++ backend/rand-chat-ws/src/main/java/com/rand/controller/ChatSocketController.java
@@ -6,7 +6,7 @@ import com.rand.chat.model.ChatMessage;
 import com.rand.common.service.PathVarValidationService;
 import com.rand.config.constant.PubSubChannel;
 import com.rand.constant.ChatConst;
-import com.rand.service.ChatWebFluxService;
+import com.rand.service.ChatIOService;
 import com.rand.util.ChatUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,15 +17,16 @@ import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.stereotype.Controller;
-
+//메시지 컨트롤러
 @Controller
 @RequiredArgsConstructor
 @Slf4j
 public class ChatSocketController {
 
     private final RedisTemplate redisTemplate;
-    private final ChatWebFluxService chatWebFluxService;
-//채팅 stomp 통신
+    private final ChatIOService chatIOService;
+
+    //채팅 stomp 통신
     @MessageMapping("/room/{roomId}")
             public void test(@Payload ReqChatMsgDTO message,
                              @DestinationVariable int roomId,
@@ -39,8 +40,8 @@ public class ChatSocketController {
        }
         ChatUtil.concatMessage(incomingMessage,chatMessage,roomId, ChatConst.PUB_CHAT_ROOM_URL);
 
-       //메시지 전송시 상대방과 내가 해당방에 참여하는지 체크 후 읽음여부를 업데이트 및 웹소켓으로 읽음 플래그를 전송
-        chatWebFluxService.updateIsReadOfSend(roomId,chatMessage.getUsrId());
+        //메시지 전송시 상대방과 내가 해당방에 참여하는지 체크 후 읽음여부를 업데이트 및 웹소켓으로 읽음 플래그를 전송
+        chatIOService.updateIsReadOfSend(roomId,chatMessage.getUsrId());
 
 
         ReqChatMsgSaveDTO reqChatMsgSaveDTO  = new ReqChatMsgSaveDTO(chatMessage);
