+++ backend/rand-chat-ws/src/main/java/com/rand/interceptor/StompInBoundInterceptor.java
@@ -4,7 +4,7 @@ import com.rand.chat.dto.request.RoomValidDTO;
 import com.rand.constant.ChatConst;
 import com.rand.jwt.JWTUtil;
 import com.rand.jwt.JwtError;
-import com.rand.service.ChatWebFluxService;
+import com.rand.service.ChatIOService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.messaging.Message;
@@ -21,9 +21,10 @@ import java.util.List;
 @Slf4j
 @Component
 @RequiredArgsConstructor
+//STOMP 통신 인터셉터
 public class StompInBoundInterceptor implements ChannelInterceptor {
     private final JWTUtil jwtUtil;
-    private final ChatWebFluxService chatWebFluxService;
+    private final ChatIOService chatWebFluxService;
 
 //메시지 전송전 실행되어 핸들링
     @Override
