+++ backend/rand-chat-ws/src/main/java/com/rand/service/ChatIOService.java
@@ -4,12 +4,11 @@ import com.rand.chat.dto.request.RoomValidDTO;
 import com.rand.common.ResponseDTO;
 import com.rand.member.model.Members;
 import org.springframework.http.ResponseEntity;
-import org.springframework.messaging.handler.annotation.Header;
 import reactor.core.publisher.Mono;
 
-public interface ChatWebFluxService {
-    public Mono<Members> checkEnterUser(Integer chatRoomId, String accessToken);
+public interface ChatIOService {
+
     public void updateIsReadOfSend(Integer chatRoomId,String accessToken);
     public Boolean isRealYourChatRoom(RoomValidDTO roomValidDTO,String accessToken);
-    public ResponseEntity<ResponseDTO<Void>> enterRoomUpdateInfo(Integer chatRoomId, String accessToken);
+    public ResponseEntity<ResponseDTO<Void>> updateIsReadOfEnter(Integer chatRoomId, String accessToken);
 }
