+++ backend/rand-chat-ws/src/main/resources/application-deploy.properties
@@ -56,4 +56,4 @@ spring.servlet.multipart.max-file-size=5MB
 # ?? ?? ?? (?? + ?? ???)
 spring.servlet.multipart.max-request-size=5MB 
 
-web.nginx.url=https://??url
+web.nginx.url=http://randchat.o-r.kr
